Move the display of code excerpts to the Report effect (draft, errors exist) #36
